refactor(transport): split assignment page into hook and presenter
